requirements.txt
Django==1.4.3
GDAL==1.9.1
Shapely==1.2.16
amqp==1.0.6
anyjson==0.3.3
billiard==2.7.3.19
celery==3.0.13
django-celery==3.0.11
django-jsonfield==0.8.12
django-uuidfield==0.4.0
importlib==1.0.2
kombu==2.5.4
numpy==1.6.2
ordereddict==1.1
pysqlite==2.6.3
python-dateutil==1.5
pytz==2012j
requests==1.0.4
rpy2==2.3.1
scipy==0.11.0
wsgiref==0.1.2
